remove commands for non existing api-keys crud endpoints, and add for api-keys-refresh endpoint
